use references to address as much as possible

+++ cortex/IODevice.h
@@ -72,7 +72,7 @@ namespace cortex
             uint32_t unixtime() noexcept;
 
             void setCompareValue(uint16_t value) noexcept;
-            uint8_t getCompareValue() noexcept;
+            uint16_t getCompareValue() noexcept;
             void setPrescalar(uint8_t value) noexcept;
             uint8_t getPrescalar() noexcept;
